Fix crash in password reset page
The `<ErrorArea />` requires the `errors` reducer. There's also a world
where we don't have the `<ErrorArea />` at all on the password reset
page, but not sure right now.
